fix: break circular import — move limiter/csrf to extensions.py (fixes mypy CI) (#42)
* fix: null-guard paginationEl IIFE and renderPaginationBar for viewer role

findingsPagination lives inside the admin/auditor-only Jinja2 block, so
getElementById returns null for viewers. The IIFE at page load called
.addEventListener directly on null, crashing immediately.

Fix: wrap IIFE body in if (paginationEl). Also add early return to
renderPaginationBar when el is null to prevent a latent crash if the
function is ever reached without the element present.


* fix: skip loadSettings() fetch for non-admin roles to suppress 403 noise

GET /settings returns 403 for auditor and viewer. The call was unconditional
so every non-admin login logged a 403 in the browser console. JS-side
appSettings defaults (auto_pdf:false, auto_archive:false, compliance:"") are
sensible for non-admin roles; no request needed.


* fix: respect WEB_CONCURRENCY on Render; fix HEALTHCHECK port; add .dockerignore

Root cause of Render auto-deploy failures (exited with status 3):
gunicorn.conf.py read GUNICORN_WORKERS (default 2) instead of Render's
WEB_CONCURRENCY=1. On a 1-CPU instance, starting 2 workers during a
zero-downtime deploy (old container still live) caused memory pressure and
OOM kills on the new workers. Gunicorn exits with code 3 when workers
repeatedly fail to boot. Clear-build-cache deploys worked because the longer
build time allowed the old container to stop before new workers loaded.

Fix: read WEB_CONCURRENCY first (Render/PaaS), fall back to GUNICORN_WORKERS
(self-hosted override), then default to 2.

Also fixes Docker HEALTHCHECK hardcoded to port 5000 while Render sets
PORT=10000 — healthcheck now uses ${PORT:-5000} and adds 5s extra
start-period to give workers time to boot before first check.

Adds .dockerignore to exclude .git, .claude/ worktrees, tests/, docs/,
__pycache__, and other dev artifacts from the Docker image.


* style: ruff format — fix formatting across Python files


* fix: break circular import by moving limiter/csrf to extensions.py

audit, history, reports, and api_v1 blueprints imported limiter/csrf back
from cashel.web, creating a circular import that mypy couldn't type-check
(web.py → blueprint → cashel.web → blueprint). This caused [has-type] errors
on the register_blueprint() calls in web.py.

Fix: introduce src/cashel/extensions.py holding uninitialized CSRFProtect and
Limiter instances. web.py calls .init_app(app) on each after creating the Flask
app. All four blueprints now import from cashel.extensions, not cashel.web.

Verified: mypy clean, ruff lint clean, ruff format clean.


---------
